voice_mood_app/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
librosa==0.10.1
numpy==1.24.3
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
django-cors-headers==4.3.1